Separates Storage and InternalStorage.
